requirements.txt
fastapi
uvicorn[standard]
mysql-connector-python
tabula-py
tabulate
